feat: Refactor daily forecast and add hourly forecast feature
This commit introduces a major refactor of the daily forecast feature and implements the foundational components for the new hourly forecast feature.

Key changes include:
- **Renamed Domain and Entity Models**:
    - `Daily` is now `DayForecastSimple`.
    - `DailyEntity` is now `DayForecastEntity` and `DayForecastSimpleEntity`.
    - `DailyDao` is now `DayForecastDao`.
    - The `core:data` module's directory structure has been reorganized to group database and remote data source files.
- **Enhanced Database Entities**:
    - `DayForecastEntity` now includes `sunrise`, `sunset`, and an embedded `Current` object for current weather conditions.
    - New entities `HourForecastEntity`, `DayWithHours`, and `Current` have been created to support detailed and hourly forecasts.
- **New Hourly Forecast Feature**:
    - Created the `feature:forecast_hourly` module with `data`, `domain`, `di`, and `presentation` layers.
    - Added domain models: `DayForecastDetail`, `HourForecast`, and `Alert`.
    - Implemented `ForecastDayDetailRepository` and `ForecastHourlyRemoteDataSource` to fetch detailed daily and hourly data.
    - New DTOs (`ForecastDayDetailDto`, `CurrentDto`, `HourlyDto`, etc.) have been added to handle the API response for detailed forecasts.
- **Updated Database and DI**:
    - `AppDataBase` is updated to include `HourForecastDao` and `HourForecastEntity`.
    - A new `forecastHourlyModule` has been created for Koin dependency injection and integrated into the application.
- **Refined Data Handling**:
    - Introduced a generic `networkBoundResource` function to handle single-object data streams alongside the existing list-based one.
    - Updated mappers and repositories to align with the new data models and DTOs.
    - Added extension functions for `Long` to `LocalDateTime` and `toStartOfDayUtc` conversions.

+++ feature/forecast_daily/data/src/main/java/com/panabuntu/weathertracker/feature/forecast_daily/data/remote_data_source/dto/DailyDto.kt
@@ -1,6 +1,7 @@
 package com.panabuntu.weathertracker.feature.forecast_daily.data.remote_data_source.dto
 
 
+import com.panabuntu.weathertracker.core.data.remote_data_source.dto.WeatherDto
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
@@ -31,9 +32,9 @@ data class DailyDto(
     @SerialName("summary")
     val summary: String,
     @SerialName("sunrise")
-    val sunrise: Int,
+    val sunrise: Long,
     @SerialName("sunset")
-    val sunset: Int,
+    val sunset: Long,
     @SerialName("temp")
     val temp: TempDto,
     @SerialName("uvi")
